+++ app/src/main/java/com/github/se/polyfit/ui/progressTracker/calTracker.kt
@@ -12,7 +12,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.github.se.polyfit.ui.utils.OverviewTags
@@ -51,10 +50,4 @@ fun CalTracker(
                     fontSize = 18.sp, color = color, fontWeight = FontWeight.Bold),
             modifier = Modifier.align(Alignment.Center).testTag(OverviewTags.overviewCalorie))
       }
-}
-
-@Composable
-@Preview
-fun CalTrackerPreview() {
-  CalTracker(.5f, "something")
 }
