+++ app/src/main/java/com/github/se/polyfit/viewmodel/meal/OverviewViewModel.kt
@@ -1,7 +1,15 @@
 package com.github.se.polyfit.viewmodel.meal
 
+import android.Manifest
+import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
 import android.graphics.Bitmap
+import android.provider.MediaStore
 import android.util.Log
+import androidx.activity.compose.ManagedActivityResultLauncher
+import androidx.activity.result.ActivityResult
+import androidx.core.content.ContextCompat
 import androidx.lifecycle.ViewModel
 import com.github.se.polyfit.data.api.SpoonacularApiCaller
 import com.github.se.polyfit.data.local.dao.MealDao
@@ -171,10 +179,33 @@ constructor(
   }
 
   fun getUserName(): String {
-    return if (user.displayName != null) {
-      user.displayName!!
-    } else {
-      user.email
+    return when {
+      user.displayName != null -> user.displayName!!
+      else -> user.email
+    }
+  }
+
+  fun callCamera(
+      context: Context,
+      startCamera: ManagedActivityResultLauncher<Intent, ActivityResult>,
+      requestPermissionLauncher: ManagedActivityResultLauncher<String, Boolean>
+  ): () -> Unit = {
+    // Check if the permission has already been granted
+    when (ContextCompat.checkSelfPermission(context, Manifest.permission.CAMERA)) {
+      PackageManager.PERMISSION_GRANTED -> {
+        // You can use the camera
+        val takePictureIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+        try {
+          startCamera.launch(takePictureIntent)
+        } catch (e: Exception) {
+          // Handle the exception if the camera intent cannot be launched
+          Log.e("HomeScreen", "Error launching camera intent: $e")
+        }
+      }
+      else -> {
+        // Request the permission
+        requestPermissionLauncher.launch(Manifest.permission.CAMERA)
+      }
     }
   }
 }
